Split project into frontend and backend with updated structure, env vars, and config
