+++ frontend/next.config.ts
