+++ frontend/package.json
@@ -1,19 +1,17 @@
 {
-  "name": "shared-story",
+  "name": "shared-story-frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "node server.mjs",
+    "dev": "next dev",
     "build": "next build",
-    "start": "NODE_ENV=production node server.mjs"
+    "start": "next start"
   },
   "dependencies": {
     "@heroicons/react": "^2.2.0",
-    "express": "^5.1.0",
     "next": "15.3.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1"
   },
   "devDependencies": {
